put struct Backend in backends module, create BackendError
