+++ lib/src/backends.rs
@@ -1,16 +1,182 @@
 use std::{cell::RefCell, collections::HashMap, net::SocketAddr, rc::Rc};
 
-use anyhow::{bail, Context};
 use mio::net::TcpStream;
 
 use sozu_command::{
-    proto::command::{Event, EventKind, LoadBalancingAlgorithms, LoadMetric},
+    proto::command::{Event, EventKind, LoadBalancingAlgorithms, LoadBalancingParams, LoadMetric},
     state::ClusterId,
 };
+use time::Duration;
 
-use crate::server::push_event;
+use crate::{
+    retry::{self, RetryPolicy},
+    server::{self, push_event},
+    PeakEWMA,
+};
 
-use super::{load_balancing::*, Backend};
+use super::load_balancing::*;
+
+#[derive(thiserror::Error, Debug)]
+pub enum BackendError {
+    #[error("No backend found for cluster {0}")]
+    NoBackendForCluster(String),
+    #[error("No more backend is available for cluster {0}")]
+    NoAvailableBackendForCluster(String),
+    #[error("Failed to connect to socket with MIO: {0}")]
+    MioConnectionError(String),
+    #[error("This backend is not in a normal status: status={0:?}")]
+    StatusNotNormal(BackendStatus),
+    #[error(
+        "could not connect {cluster_id} to {backend_address:?} ({failures} failures): {error}"
+    )]
+    ConnectionFailures {
+        cluster_id: String,
+        backend_address: SocketAddr,
+        failures: usize,
+        error: String,
+    },
+}
+
+#[derive(Debug, PartialEq, Eq, Clone)]
+pub enum BackendStatus {
+    Normal,
+    Closing,
+    Closed,
+}
+
+#[derive(Debug, PartialEq, Clone)]
+pub struct Backend {
+    pub sticky_id: Option<String>,
+    pub backend_id: String,
+    pub address: SocketAddr,
+    pub status: BackendStatus,
+    pub retry_policy: retry::RetryPolicyWrapper,
+    pub active_connections: usize,
+    pub active_requests: usize,
+    pub failures: usize,
+    pub load_balancing_parameters: Option<LoadBalancingParams>,
+    pub backup: bool,
+    pub connection_time: PeakEWMA,
+}
+
+impl Backend {
+    pub fn new(
+        backend_id: &str,
+        address: SocketAddr,
+        sticky_id: Option<String>,
+        load_balancing_parameters: Option<LoadBalancingParams>,
+        backup: Option<bool>,
+    ) -> Backend {
+        let desired_policy = retry::ExponentialBackoffPolicy::new(6);
+        Backend {
+            sticky_id,
+            backend_id: backend_id.to_string(),
+            address,
+            status: BackendStatus::Normal,
+            retry_policy: desired_policy.into(),
+            active_connections: 0,
+            active_requests: 0,
+            failures: 0,
+            load_balancing_parameters,
+            backup: backup.unwrap_or(false),
+            connection_time: PeakEWMA::new(),
+        }
+    }
+
+    pub fn set_closing(&mut self) {
+        self.status = BackendStatus::Closing;
+    }
+
+    pub fn retry_policy(&mut self) -> &mut retry::RetryPolicyWrapper {
+        &mut self.retry_policy
+    }
+
+    pub fn can_open(&self) -> bool {
+        if let Some(action) = self.retry_policy.can_try() {
+            self.status == BackendStatus::Normal && action == retry::RetryAction::OKAY
+        } else {
+            false
+        }
+    }
+
+    pub fn inc_connections(&mut self) -> Option<usize> {
+        if self.status == BackendStatus::Normal {
+            self.active_connections += 1;
+            Some(self.active_connections)
+        } else {
+            None
+        }
+    }
+
+    /// TODO: normalize with saturating_sub()
+    pub fn dec_connections(&mut self) -> Option<usize> {
+        match self.status {
+            BackendStatus::Normal => {
+                if self.active_connections > 0 {
+                    self.active_connections -= 1;
+                }
+                Some(self.active_connections)
+            }
+            BackendStatus::Closed => None,
+            BackendStatus::Closing => {
+                if self.active_connections > 0 {
+                    self.active_connections -= 1;
+                }
+                if self.active_connections == 0 {
+                    self.status = BackendStatus::Closed;
+                    None
+                } else {
+                    Some(self.active_connections)
+                }
+            }
+        }
+    }
+
+    pub fn set_connection_time(&mut self, dur: Duration) {
+        self.connection_time.observe(dur.whole_nanoseconds() as f64);
+    }
+
+    pub fn peak_ewma_connection(&mut self) -> f64 {
+        self.connection_time.get(self.active_connections)
+    }
+
+    pub fn try_connect(&mut self) -> Result<mio::net::TcpStream, BackendError> {
+        if self.status != BackendStatus::Normal {
+            return Err(BackendError::StatusNotNormal(self.status.to_owned()));
+        }
+
+        match mio::net::TcpStream::connect(self.address) {
+            Ok(tcp_stream) => {
+                //self.retry_policy.succeed();
+                self.inc_connections();
+                Ok(tcp_stream)
+            }
+            Err(mio_error) => {
+                self.retry_policy.fail();
+                self.failures += 1;
+                // TODO: handle EINPROGRESS. It is difficult. It is discussed here:
+                // https://docs.rs/mio/latest/mio/net/struct.TcpStream.html#method.connect
+                // with an example code here:
+                // https://github.com/Thomasdezeeuw/heph/blob/0c4f1ab3eaf08bea1d65776528bfd6114c9f8374/src/net/tcp/stream.rs#L560-L622
+                Err(BackendError::MioConnectionError(mio_error.to_string()))
+            }
+        }
+    }
+}
+
+// when a backend has been removed from configuration and the last connection to
+// it has stopped, it will be dropped, so we can notify that the backend server
+// can be safely stopped
+impl std::ops::Drop for Backend {
+    fn drop(&mut self) {
+        server::push_event(Event {
+            kind: EventKind::RemovedBackendHasNoConnections as i32,
+            backend_id: Some(self.backend_id.clone()),
+            address: Some(self.address.to_string()),
+            cluster_id: None,
+        });
+    }
+}
 
 #[derive(Debug)]
 pub struct BackendMap {
@@ -54,7 +220,7 @@ impl BackendMap {
             .add_backend(backend);
     }
 
-    // TODO: return anyhow::Result with context, log the error downstream
+    // TODO: return <Result, BackendError>, log the error downstream
     pub fn remove_backend(&mut self, cluster_id: &str, backend_address: &SocketAddr) {
         if let Some(backends) = self.backends.get_mut(cluster_id) {
             backends.remove_backend(backend_address);
@@ -66,7 +232,7 @@ impl BackendMap {
         }
     }
 
-    // TODO: return anyhow::Result with context, log the error downstream
+    // TODO: return <Result, BackendError>, log the error downstream
     pub fn close_backend_connection(&mut self, cluster_id: &str, addr: &SocketAddr) {
         if let Some(cluster_backends) = self.backends.get_mut(cluster_id) {
             if let Some(ref mut backend) = cluster_backends.find_backend(addr) {
@@ -85,17 +251,15 @@ impl BackendMap {
     pub fn backend_from_cluster_id(
         &mut self,
         cluster_id: &str,
-    ) -> anyhow::Result<(Rc<RefCell<Backend>>, TcpStream)> {
+    ) -> Result<(Rc<RefCell<Backend>>, TcpStream), BackendError> {
         let cluster_backends = self
             .backends
             .get_mut(cluster_id)
-            .with_context(|| format!("No backend found for cluster {cluster_id}"))?;
+            .ok_or(BackendError::NoBackendForCluster(cluster_id.to_owned()))?;
 
         if cluster_backends.backends.is_empty() {
             self.available = false;
-            bail!(format!(
-                "Found an empty backend list for cluster {cluster_id}"
-            ));
+            return Err(BackendError::NoBackendForCluster(cluster_id.to_owned()));
         }
 
         let next_backend = match cluster_backends.next_available_backend() {
@@ -111,7 +275,9 @@ impl BackendMap {
                         address: None,
                     });
                 }
-                bail!("No more backend available for cluster {}", cluster_id);
+                return Err(BackendError::NoAvailableBackendForCluster(
+                    cluster_id.to_owned(),
+                ));
             }
         };
 
@@ -127,11 +293,13 @@ impl BackendMap {
             )
         );
 
-        let tcp_stream = borrowed_backend.try_connect().with_context(|| {
-            format!(
-                "could not connect {} to {:?} ({} failures)",
-                cluster_id, borrowed_backend.address, borrowed_backend.failures
-            )
+        let tcp_stream = borrowed_backend.try_connect().map_err(|backend_error| {
+            BackendError::ConnectionFailures {
+                cluster_id: cluster_id.to_owned(),
+                backend_address: borrowed_backend.address,
+                failures: borrowed_backend.failures,
+                error: backend_error.to_string(),
+            }
         })?;
         self.available = true;
 
@@ -142,7 +310,7 @@ impl BackendMap {
         &mut self,
         cluster_id: &str,
         sticky_session: &str,
-    ) -> anyhow::Result<(Rc<RefCell<Backend>>, TcpStream)> {
+    ) -> Result<(Rc<RefCell<Backend>>, TcpStream), BackendError> {
         let sticky_conn = self
             .backends
             .get_mut(cluster_id)
